# api/openapi/management.yaml
openapi: 3.0.3
info:
  title: tgwebdav Management API
  description: |
    Administrative REST API for tgwebdav. Every endpoint under /api/v1 requires
    administrator authentication (HTTP Basic with an is_admin user, or a Bearer
    token belonging to an is_admin user). The /healthz probe is public.
  version: "1.0.0"
servers:
  - url: /

paths:
  /healthz:
    get:
      operationId: healthz
      summary: Liveness probe (public, no auth)
      tags: [system]
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"

  /api/v1/users:
    get:
      operationId: listUsers
      summary: List all users
      tags: [users]
      responses:
        "200":
          description: The list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: createUser
      summary: Create a new user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: The created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      operationId: getUser
      summary: Get a single user
      tags: [users]
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      operationId: deleteUser
      summary: Delete a user
      tags: [users]
      responses:
        "204":
          description: User deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/users/{userId}/password:
    parameters:
      - $ref: "#/components/parameters/UserId"
    put:
      operationId: setUserPassword
      summary: Set a user's password
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPassword"
      responses:
        "204":
          description: Password updated
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/users/{userId}/tokens:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      operationId: listUserTokens
      summary: List a user's API tokens
      tags: [tokens]
      responses:
        "200":
          description: The list of tokens (hashes are never returned)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIToken"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: createUserToken
      summary: Create an API token for a user
      tags: [tokens]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToken"
      responses:
        "201":
          description: |
            The created token. The plaintext `token` is returned exactly once
            and is never recoverable afterwards.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedToken"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/users/{userId}/tokens/{tokenId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/TokenId"
    delete:
      operationId: deleteUserToken
      summary: Revoke an API token
      tags: [tokens]
      responses:
        "204":
          description: Token revoked
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/bots:
    get:
      operationId: listBots
      summary: List all bots
      tags: [bots]
      responses:
        "200":
          description: The list of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: addBot
      summary: Add a bot by its Telegram API token
      tags: [bots]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBot"
      responses:
        "201":
          description: The added bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/bots/{botId}:
    parameters:
      - $ref: "#/components/parameters/BotId"
    delete:
      operationId: deleteBot
      summary: Remove a bot
      tags: [bots]
      responses:
        "204":
          description: Bot removed
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/bots/{botId}/enabled:
    parameters:
      - $ref: "#/components/parameters/BotId"
    put:
      operationId: setBotEnabled
      summary: Enable or disable a bot
      tags: [bots]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetEnabled"
      responses:
        "204":
          description: Bot enabled state updated
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/channels:
    get:
      operationId: listChannels
      summary: List all channels
      tags: [channels]
      responses:
        "200":
          description: The list of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      operationId: addChannel
      summary: Add a channel by its bare numeric id
      tags: [channels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddChannel"
      responses:
        "201":
          description: The added channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/channels/{channelId}:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    delete:
      operationId: deleteChannel
      summary: Decommission a channel
      tags: [channels]
      responses:
        "204":
          description: Channel decommissioned
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/channels/{channelId}/eviction:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    put:
      operationId: setChannelEviction
      summary: Set a channel's eviction threshold
      tags: [channels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetEviction"
      responses:
        "204":
          description: Eviction threshold updated
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/settings:
    get:
      operationId: getSettings
      summary: Get runtime settings
      tags: [settings]
      responses:
        "200":
          description: The current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      operationId: updateSettings
      summary: Update runtime settings
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettings"
      responses:
        "200":
          description: The updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/stats:
    get:
      operationId: queryStats
      summary: Query a time-series metric
      tags: [stats]
      parameters:
        - name: metric
          in: query
          required: true
          description: The metric name (e.g. storage_bytes, read_bytes).
          schema:
            type: string
        - name: label
          in: query
          required: false
          description: Optional metric label/dimension.
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Inclusive lower bound of the time window (RFC 3339).
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Inclusive upper bound of the time window (RFC 3339).
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: The matching samples
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatPoint"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }

  /api/v1/events:
    get:
      operationId: listEvents
      summary: List audit/log events
      tags: [events]
      parameters:
        - name: kind
          in: query
          required: false
          description: Optional event kind filter.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of events to return (default 100).
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          required: false
          description: Number of events to skip.
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        "200":
          description: The matching events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic with an is_admin user's credentials.
    bearerAuth:
      type: http
      scheme: bearer
      description: A Management API token belonging to an is_admin user.

  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: The user's UUID.
      schema:
        type: string
        format: uuid
    TokenId:
      name: tokenId
      in: path
      required: true
      description: The API token's UUID.
      schema:
        type: string
        format: uuid
    BotId:
      name: botId
      in: path
      required: true
      description: The bot's UUID.
      schema:
        type: string
        format: uuid
    ChannelId:
      name: channelId
      in: path
      required: true
      description: The channel's UUID.
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The principal is not an administrator.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested entity does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: A unique-constraint or state conflict occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A human-readable error message.

    User:
      type: object
      required:
        - id
        - login
        - is_admin
        - quota_bytes
        - bandwidth_bps
        - rate_per_min
        - created_at
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
        is_admin:
          type: boolean
        quota_bytes:
          type: integer
          format: int64
          description: Storage quota in bytes (0 = unlimited).
        bandwidth_bps:
          type: integer
          format: int64
          description: Read bandwidth cap in bytes/sec (0 = unlimited).
        rate_per_min:
          type: integer
          format: int32
          description: Request rate cap per minute (0 = unlimited).
        created_at:
          type: string
          format: date-time

    CreateUser:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        is_admin:
          type: boolean
          default: false
        quota_bytes:
          type: integer
          format: int64
          default: 0
        bandwidth_bps:
          type: integer
          format: int64
          default: 0
        rate_per_min:
          type: integer
          format: int32
          default: 0

    SetPassword:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1

    APIToken:
      type: object
      required:
        - id
        - user_id
        - name
        - created_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true

    CreateToken:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    CreatedToken:
      type: object
      required:
        - id
        - user_id
        - name
        - token
        - created_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        token:
          type: string
          description: The plaintext bearer token, returned exactly once.
        created_at:
          type: string
          format: date-time

    Bot:
      type: object
      required:
        - id
        - username
        - enabled
        - created_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        enabled:
          type: boolean
        unavailable_until:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    AddBot:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
          description: The Telegram Bot API token.

    SetEnabled:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean

    Channel:
      type: object
      required:
        - id
        - tg_chat_id
        - title
        - message_counter
        - eviction_threshold
        - available
        - created_at
      properties:
        id:
          type: string
          format: uuid
        tg_chat_id:
          type: integer
          format: int64
        title:
          type: string
        message_counter:
          type: integer
          format: int64
        eviction_threshold:
          type: integer
          format: int64
        available:
          type: boolean
        created_at:
          type: string
          format: date-time

    AddChannel:
      type: object
      required: [bare_id]
      properties:
        bare_id:
          type: integer
          format: int64
          description: The bare channel id (the -100 prefix is applied internally).

    SetEviction:
      type: object
      required: [threshold]
      properties:
        threshold:
          type: integer
          format: int64
          minimum: 0

    Settings:
      type: object
      required:
        - blob_max_size
        - wal_idle_timeout_ms
        - max_file_size
        - default_eviction_threshold
        - updated_at
      properties:
        blob_max_size:
          type: integer
          format: int64
          description: Working blob size in bytes.
        wal_idle_timeout_ms:
          type: integer
          format: int64
          description: WAL idle flush timeout in milliseconds.
        max_file_size:
          type: integer
          format: int64
          description: Maximum single uploaded file size in bytes (0 = unlimited).
        default_eviction_threshold:
          type: integer
          format: int64
          description: Eviction threshold seeded onto new channels.
        updated_at:
          type: string
          format: date-time

    UpdateSettings:
      type: object
      properties:
        blob_max_size:
          type: integer
          format: int64
        wal_idle_timeout_ms:
          type: integer
          format: int64
        max_file_size:
          type: integer
          format: int64
        default_eviction_threshold:
          type: integer
          format: int64

    StatPoint:
      type: object
      required:
        - ts
        - metric
        - label
        - value
      properties:
        ts:
          type: string
          format: date-time
        metric:
          type: string
        label:
          type: string
        value:
          type: number
          format: double

    Event:
      type: object
      required:
        - id
        - ts
        - kind
        - message
        - ref
      properties:
        id:
          type: string
          format: uuid
        ts:
          type: string
          format: date-time
        kind:
          type: string
        message:
          type: string
        ref:
          type: string

    EventPage:
      type: object
      required:
        - events
        - total
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
          format: int64

security:
  - basicAuth: []
  - bearerAuth: []